fix: dotnet runtime dependency-mismatch
Since the original tools were build in a ubuntu:focal-container and
copied into the Microsoft one (based on some old Debian), there was a
mismatch in runtime libraries.
This should be fixed by relying on the Microsoft-Image only, by building
the ilasm-Assembler within it.
In order to keep the Image small, most build-artifacts and build-tools
are discarded by copying only the relevant stuff into the final Image.